Added SpriteAnimation namespace to content pipeline. Renamed ContentProcessor1 to SpriteAnimationProcessor and moved it to SpriteAnimation namespace. Renamed ContentImporter1 to SpriteAnimationImporter and moved it to SpriteAnimation namespace.
