+++ Yetibyte.Himalaya.Content.Pipeline/SpriteAnimation/SpriteAnimationProcessor.cs
@@ -4,7 +4,7 @@ using Microsoft.Xna.Framework.Content.Pipeline;
 using TInput = System.String;
 using TOutput = System.String;
 
-namespace Yetibyte.Himalaya.Content.Pipeline {
+namespace Yetibyte.Himalaya.Content.Pipeline.SpriteAnimation {
     /// <summary>
     /// This class will be instantiated by the XNA Framework Content Pipeline
     /// to apply custom processing to content data, converting an object of
@@ -16,8 +16,8 @@ namespace Yetibyte.Himalaya.Content.Pipeline {
     /// TODO: change the ContentProcessor attribute to specify the correct
     /// display name for this processor.
     /// </summary>
-    [ContentProcessor(DisplayName = "Yetibyte.Himalaya.Content.Pipeline.ContentProcessor1")]
-    public class ContentProcessor1 : ContentProcessor<TInput, TOutput> {
+    [ContentProcessor(DisplayName = "Himalaya Sprite Animation Processor")]
+    public class SpriteAnimationProcessor : ContentProcessor<TInput, TOutput> {
         public override TOutput Process(TInput input, ContentProcessorContext context) {
             // TODO: process the input object, and return the modified data.
             throw new NotImplementedException();
