+++ Yetibyte.Himalaya.Content.Pipeline/Yetibyte.Himalaya.Content.Pipeline.csproj
@@ -47,8 +47,8 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ContentImporter1.cs" />
-    <Compile Include="ContentProcessor1.cs" />
+    <Compile Include="SpriteAnimation\SpriteAnimationImporter.cs" />
+    <Compile Include="SpriteAnimation\SpriteAnimationProcessor.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
